LayoutTests/http/tests/security/contentSecurityPolicy/directive-name-case-insensitive.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="Frame-Ancestors 'none'; Report-URI /report; Aaa">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<p>This tests that mixed-case Content Security Policy directive names are lowercased, so the console warnings refer to them in lowercase.</p>
</body>
</html>
